Выполнила задание module5-task1
